// apps/web/components/ui/select.d.ts
import React from 'react';
export declare const Select: ({ children, value, onValueChange, className }: {
    children: React.ReactNode;
    value?: string;
    onValueChange?: (value: string) => void;
    className?: string;
}) => React.JSX.Element;
export declare const SelectTrigger: ({ children, className }: {
    children: React.ReactNode;
    className?: string;
}) => React.JSX.Element;
export declare const SelectContent: ({ children }: {
    children: React.ReactNode;
}) => React.JSX.Element;
export declare const SelectItem: ({ children, value }: {
    children: React.ReactNode;
    value: string;
}) => React.JSX.Element;
export declare const SelectValue: ({ placeholder }: {
    placeholder?: string;
}) => React.JSX.Element;
//# sourceMappingURL=select.d.ts.map