// apps/web/components/ui/tabs.d.ts
import React from 'react';
export declare const Tabs: ({ children, defaultValue, className }: {
    children: React.ReactNode;
    defaultValue?: string;
    className?: string;
}) => React.JSX.Element;
export declare const TabsList: ({ children, className }: {
    children: React.ReactNode;
    className?: string;
}) => React.JSX.Element;
export declare const TabsTrigger: ({ children, value, className }: {
    children: React.ReactNode;
    value: string;
    className?: string;
}) => React.JSX.Element;
export declare const TabsContent: ({ children, value, className }: {
    children: React.ReactNode;
    value: string;
    className?: string;
}) => React.JSX.Element;
//# sourceMappingURL=tabs.d.ts.map